Revamp of dat library structure, Xpress model.  Still work to do

+++ reo/library.py
@@ -1,5 +1,7 @@
 import os
 import subprocess
+import traceback
+
 import utilities
 import pandas as pd
 
@@ -7,7 +9,7 @@ import pandas as pd
 class dat_library:
 
     # if need to debug, change to True, outputs OUT files, GO files, debugging to cmdline
-    debug = False
+    debug = True
 
     run_id = []
     run_file = []
@@ -39,11 +41,13 @@ class dat_library:
     # directory structure
     path_xpress = []
     path_dat_library = []
+    path_various = []
     path_load_size = []
+    path_load_profile = []
+    path_economics = []
     path_pv_om = []
     path_batt_cost_kwh = []
     path_batt_cost_kw = []
-    path_load_profile = []
     path_pv_cost = []
     path_owner_discount_rate = []
     path_offtaker_discount_rate = []
@@ -57,6 +61,7 @@ class dat_library:
     def __init__(self, run_id, path_xpress, analysis_period, latitude, longitude, load_size, pv_om, batt_cost_kw,
                  batt_cost_kwh, load_profile, pv_cost, owner_discount_rate, offtaker_discount_rate,
                  utility_name, rate_name):
+
         self.path_xpress = path_xpress
 
         self.run_id = run_id
@@ -67,10 +72,10 @@ class dat_library:
         self.latitude = latitude
         self.longitude = longitude
         self.load_size = load_size
+        self.load_profile = load_profile
         self.pv_om = pv_om
         self.batt_cost_kw = batt_cost_kw
         self.batt_cost_kwh = batt_cost_kwh
-        self.load_profile = load_profile
         self.pv_cost = pv_cost
         self.owner_discount_rate = owner_discount_rate
         self.offtaker_discount_rate = offtaker_discount_rate
@@ -87,7 +92,13 @@ class dat_library:
         self.create_or_load()
         self.create_run_file()
 
+        #print ('New subprocess')
+
+        #tracefile = open('traceback.txt', 'a')
+        #traceback.print_stack(limit=5, file=tracefile)
         subprocess.call(self.run_file)
+        # print ('Subprocess done')
+
         self.parse_outputs()
         self.cleanup()
 
@@ -96,18 +107,21 @@ class dat_library:
     def define_paths(self):
 
         self.path_dat_library = os.path.join(self.path_xpress, "DatLibrary")
+        self.path_various = os.path.join(self.path_dat_library, "Various")
+        self.path_economics = os.path.join(self.path_dat_library, "Economics")
         self.path_load_size = os.path.join(self.path_dat_library, "LoadSize")
+        self.path_load_profile = os.path.join(self.path_dat_library, "LoadProfiles")
+        self.path_solar_resource = os.path.join(self.path_dat_library, "GISdata")
+        self.path_utility_rate = os.path.join(self.path_dat_library, "Utility", "Los Angeles Department of Water & Power")
+        self.path_output = os.path.join(self.path_dat_library, "Output")
+
+        # Going away
         self.path_pv_om = os.path.join(self.path_dat_library, "OM")
         self.path_batt_cost_kwh = os.path.join(self.path_dat_library, "BatteryCost", "KWH")
         self.path_batt_cost_kw = os.path.join(self.path_dat_library, "BatteryCost", "KW")
-        self.path_load_profile = os.path.join(self.path_dat_library, "LoadProfiles")
         self.path_pv_cost = os.path.join(self.path_dat_library, "PVcost")
         self.path_owner_discount_rate = os.path.join(self.path_dat_library, "DiscountRates", "Owner")
         self.path_offtaker_discount_rate = os.path.join(self.path_dat_library, "DiscountRates", "Offtaker")
-        self.path_solar_resource = os.path.join(self.path_dat_library, "SolarResource")
-        self.path_utility_rate = os.path.join(self.path_dat_library, "Utility")
-
-        self.path_output = os.path.join(self.path_dat_library, "Output")
 
     def create_or_load(self):
         if self.load_size and self.load_size > 0:
@@ -152,22 +166,25 @@ class dat_library:
 
     def parse_outputs(self):
 
-        df = pd.read_csv(self.output_file, header=None, index_col=0)
-        df = df.transpose()
+        if os.path.exists(self.output_file):
+            df = pd.read_csv(self.output_file, header=None, index_col=0)
+            df = df.transpose()
 
-        if 'LCC' in df.columns:
-            self.outputs['lcc'] = df['LCC']
-        if 'Batt size KW' in df.columns:
-            self.outputs['batt_size_kw'] = df['Batt size KW']
-        if 'Batt size KWH' in df.columns:
-            self.outputs['batt_size_kwh'] = df['Batt size KWH']
-        if 'PVNM size KW' in df.columns:
-            self.outputs['pv_kw'] = df['PVNM size KW']
+            if 'LCC' in df.columns:
+                self.outputs['lcc'] = df['LCC']
+            if 'Batt size KW' in df.columns:
+                self.outputs['batt_size_kw'] = df['Batt size KW']
+            if 'Batt size KWH' in df.columns:
+                self.outputs['batt_size_kwh'] = df['Batt size KWH']
+            if 'PVNM size KW' in df.columns:
+                self.outputs['pv_kw'] = df['PVNM size KW']
 
     def cleanup(self):
         if not self.debug:
-            os.remove(self.output_file)
-            os.remove(self.run_file)
+            if os.path.exists(self.output_file):
+                os.remove(self.output_file)
+            if os.path.exists(self.run_file):
+                os.remove(self.run_file)
 
     def write_var(self, f, var, dat_var):
         f.write(dat_var + ": [\n")
@@ -193,17 +210,42 @@ class dat_library:
             self.write_var(f, var2, dat_var2)
             f.close()
 
-    # DAT1 - LoadSize
+    # DAT1 - Constant
+    # DAT2 - Economics
+
+    # DAT3 - LoadSize
     def create_load_size(self):
         path = self.path_load_size
         var = self.load_size
         filename = "LoadSize_" + var + ".dat"
         dat_var = "AnnualElecLoad"
-        self.DAT[0] = "DAT1=" + "'" + os.path.join(path, filename) + "'"
+        self.DAT[2] = "DAT3=" + "'" + os.path.join(path, filename) + "'"
 
         self.write_single_variable(path, filename, var, dat_var)
 
+    # DAT4 - Load
+    def create_load_profile(self):
+        path = self.path_load_profile
+        var = self.load_profile
+        if var.lower() in self.default_load_profiles:
+            filename = "Load8760_" + var + ".dat"
+            self.DAT[3] = "DAT4=" + "'" + os.path.join(path, filename) + "'"
 
+    # DAT5 - Solar Resource
+    # DAT6 - Storage
+    # DAT7 - Max Sizes
+
+    # DAT8 - DAT11 - Utility Rates
+    # Currently depends on files being present in directory structure
+    def create_utility_rate(self):
+        rate_path = os.path.join(self.path_utility_rate, self.utility_name, self.rate_name)
+        self.DAT[7] = "DAT8=" + "'" + os.path.join(rate_path, "TimeStepsDemand.dat")
+        self.DAT[8] = "DAT9=" + "'" + os.path.join(rate_path, "DemandRate.dat")
+        self.DAT[9] = "DAT10=" + "'" + os.path.join(rate_path, "FuelCost.dat")
+        self.DAT[10] = "DAT11=" + "'" + os.path.join(rate_path, "Export.dat")
+
+
+    '''
     # DAT2 - PVOM
     def create_pv_om(self):
         path = self.path_pv_om
@@ -234,13 +276,6 @@ class dat_library:
 
         self.write_single_variable(path, filename, var, dat_var)
 
-    # DAT5 - Load
-    def create_load_profile(self):
-        path = self.path_load_profile
-        var = self.load_profile
-        if var.lower() in self.default_load_profiles:
-            filename = "Load8760_" + var + ".dat"
-            self.DAT[4] = "DAT5=" + "'" + os.path.join(path, filename) + "'"
 
     # DAT6 - PVCost
     def create_pv_cost(self):
@@ -252,14 +287,6 @@ class dat_library:
 
         self.write_single_variable(path, filename, [var, 0], dat_var)
 
-    # DAT12 - DAT15 - Utility Rates
-    # Currently depends on files being present in directory structure
-    def create_utility_rate(self):
-        rate_path = os.path.join(self.path_utility_rate, self.utility_name, self.rate_name)
-        self.DAT[11] = "DAT12=" + "'" + os.path.join(rate_path, "DPeriods.dat")
-        self.DAT[12] = "DAT13=" + "'" + os.path.join(rate_path, "DRate.dat")
-        self.DAT[13] = "DAT14=" + "'" + os.path.join(rate_path, "ERate.dat")
-        self.DAT[14] = "DAT15=" + "'" + os.path.join(rate_path, "Export.dat")
 
     # DAT19 - Owner Discount Rate
     def create_owner_discount_rate(self):
@@ -305,5 +332,6 @@ class dat_library:
     # Utility rates, hookup urdb processor
 
     # Calculate PWFs from discount rate, or use defaults
+    '''
